requirements.txt
# Core dependencies
flask==3.1.0
flask-socketio==5.5.1
flask-cors==5.0.1
python-dotenv==1.0.1

# Browser automation
playwright==1.55.0

# AI/ML
google-generativeai==0.8.5
anthropic==0.72.0

# Utilities
python-engineio==4.12.3
python-socketio==5.14.1
Pillow==11.1.0